Add files via upload
Added four files, including a Python code file, two data files, and a file generated by running the code. These files are designed for estimating the Yield to Maturity (YTM) of a five-year TIPS. In comparison to model1, this modification introduces additional variables, such as CPI, aiming to enhance the accuracy of the estimation.